productos / categorias

+++ back/scripts/procedures/spGestionCategorias.sql
@@ -10,7 +10,9 @@ CREATE PROCEDURE spGestionCategorias(
     IN p_userID INT,
     IN p_nombre VARCHAR(50),
     IN p_descripcion VARCHAR(300),
-    IN p_categoryImage VARCHAR(200)
+    IN p_categoryImage VARCHAR(200),
+     IN p_Limit INT, 
+    IN p_Offset INT
 )
 BEGIN
     IF p_Accion = 'IN' THEN
@@ -21,18 +23,31 @@ BEGIN
     IF p_Accion = 'UP' THEN
         UPDATE Categoria
         SET nombre = p_nombre,
-            descripcion = p_descripcion
-        WHERE categoryID = p_categoryID;
+            descripcion = p_descripcion,
+            categoryImage = p_categoryImage
+        WHERE categoryID = p_categoryID and userID = p_userID;
     END IF;
 
     IF p_Accion = 'BO' THEN
-        DELETE FROM Categoria WHERE categoryID = p_categoryID;
+        DELETE FROM Categoria WHERE categoryID = p_categoryID and userID = p_userID;
     END IF;
 
     IF p_Accion = 'SE' THEN
         SELECT categoryID, nombre, descripcion,categoryImage
         FROM Categoria
         WHERE userID = p_userID;
+    END IF;
+    
+    IF p_Accion = 'SE2' THEN
+        SELECT categoryID, nombre, descripcion
+        FROM Categoria
+        ;
+    END IF;
+    
+    IF p_Accion = 'SE3' THEN
+        SELECT categoryID, nombre, descripcion,categoryImage,total
+        FROM viCategorias
+        LIMIT p_Limit OFFSET p_Offset;
     END IF;
 END //
